VinVideo_Connected/requirements-concurrent.txt
# Python dependencies for concurrent RunPod image generation
# Core dependencies for concurrent processing

aiohttp>=3.8.0          # Async HTTP client for Phase 2 implementation
requests>=2.28.0        # HTTP client for Phase 1 and health checks
python-dotenv>=0.19.0   # Environment variable loading

# Built-in Python libraries (no installation needed):
# - asyncio (Python 3.7+)
# - concurrent.futures (Python 3.2+)
# - threading (Python 3.x)
# - json (Python 3.x)
# - base64 (Python 3.x)
# - pathlib (Python 3.4+)
# - dataclasses (Python 3.7+)
# - typing (Python 3.5+)
# - logging (Python 3.x)

# Optional performance enhancements:
# uvloop>=0.17.0        # Faster event loop for asyncio (Linux/macOS only)
# orjson>=3.8.0         # Faster JSON parsing (optional)